+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/components/ResourcePicker/NestedRow.test.tsx
@@ -1,0 +1,21 @@
+import { render, screen } from '@testing-library/react';
+import React from 'react';
+import NestedRow from './NestedRow';
+import { ResourceRowType } from './types';
+
+const defaultProps = {
+  row: [],
+  level: 0,
+  selectedRows: [],
+  requestNestedRows: jest.fn(),
+  onRowSelectedChange: jest.fn(),
+  selectableEntryTypes: [],
+};
+
+describe('NestedRow', () => {
+  it('should be selectable', () => {
+    render(<NestedRow {...defaultProps} />);
+    const box = screen.getByRole('checkbox');
+    expect(box).toBeInTheDocument();
+  });
+});
